+++ tests/metrics/test_primitives.py
@@ -4,7 +4,7 @@ import torch
 from torch.autograd import Variable
 
 import torchbearer
-from torchbearer.metrics import Loss, Epoch, CategoricalAccuracy, TopKCategoricalAccuracy, BinaryAccuracy
+from torchbearer.metrics import Loss, Epoch, CategoricalAccuracy, TopKCategoricalAccuracy, BinaryAccuracy, MeanSquaredError
 
 
 class TestLoss(unittest.TestCase):
@@ -162,5 +162,35 @@ class TestTopKCategoricalAccuracy(unittest.TestCase):
         result = self._metric.process(self._state)
         for i in range(0, len(self._targets)):
             self.assertEqual(result[i], self._targets[i],
+                             msg='returned: ' + str(result[i]) + ' expected: ' + str(self._targets[i])
+                                 + ' in: ' + str(result))
+
+
+class TestMeanSquaredError(unittest.TestCase):
+    def setUp(self):
+        self._state = {
+            torchbearer.Y_TRUE: Variable(torch.FloatTensor(
+                [0.8, 0.2, 0.0, 0.4, 0.3, 0.7]
+            )),
+            torchbearer.Y_PRED: Variable(torch.FloatTensor(
+                [0.9, 0.1, 0.1, 0.7, 0.5, 0.6]
+            ))
+        }
+        self._targets = [0.01, 0.01, 0.01, 0.09, 0.04, 0.01]
+        self._metric = MeanSquaredError().root  # Get root node of Tree for testing
+
+    def test_train_process(self):
+        self._metric.train()
+        result = self._metric.process(self._state)
+        for i in range(0, len(self._targets)):
+            self.assertAlmostEqual(result[i].item(), self._targets[i], places=3,
+                             msg='returned: ' + str(result[i]) + ' expected: ' + str(self._targets[i])
+                                 + ' in: ' + str(result))
+
+    def test_validate_process(self):
+        self._metric.eval()
+        result = self._metric.process(self._state)
+        for i in range(0, len(self._targets)):
+            self.assertAlmostEqual(result[i].item(), self._targets[i], places=3,
                              msg='returned: ' + str(result[i]) + ' expected: ' + str(self._targets[i])
                                  + ' in: ' + str(result))